chore: separate helper
